reflector/requirements.txt
websockets
pillow 